pyproject.toml
[project]
name = "money"
version = "0.1.0"
description = "Personal finance data collection and analysis"
readme = "README.md"
authors = [
    { name = "Scott Kirklin", email = "scott.kirklin@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "camoufox>=0.4.11",
    "click",
    "google-cloud-storage>=3.9.0",
    "playwright",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "ruff",
    "pyright",
]
gcs = [
    "google-cloud-storage",
]

[project.scripts]
money = "money.cli:main"

[build-system]
requires = ["uv_build>=0.10.9,<0.11.0"]
build-backend = "uv_build"

[tool.ruff]
line-length = 100
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.pyright]
pythonVersion = "3.12"
typeCheckingMode = "strict"
venvPath = "."
venv = ".venv"
reportUnknownMemberType = "warning"
reportUnknownArgumentType = "warning"

[tool.pytest.ini_options]
testpaths = ["tests"]
